+++ components/social-farcaster/proposal/proposal-status-actions/MintAction.tsx
@@ -44,7 +44,7 @@ export function ReadyToMintButton({
           tokenId: Number(proposal.tokenId),
         }}
         cast={cast}
-        renderButton={(props) => {
+        renderButton={(props:any) => {
           return (
             <ProposalButton variant={"ready-to-mint"} {...props}>
               <DiamondPlus className="size-4 stroke-proposalReadyToMint-foreground" />
@@ -52,7 +52,7 @@ export function ReadyToMintButton({
             </ProposalButton>
           );
         }}
-        onSuccess={(newMintNFTAmount) => {
+        onSuccess={(newMintNFTAmount:number) => {
           // console.log("newMintNFTAmount", newMintNFTAmount, proposal);
           upsertOneToProposals(cast.hash as `0x${string}`, {
             ...proposal,
